+++ atlasdb-api/src/main/java/com/palantir/atlasdb/transaction/api/expectations/ExpectationsStatistics.java
@@ -18,7 +18,6 @@ package com.palantir.atlasdb.transaction.api.expectations;
 
 import com.google.common.collect.ImmutableMap;
 import com.palantir.atlasdb.keyvalue.api.TableReference;
-import com.palantir.atlasdb.transaction.api.TransactionReadInfo;
 import org.immutables.value.Value;
 
 @Value.Immutable
