+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/transaction/impl/TrackingKeyValueService.java
@@ -19,7 +19,7 @@ package com.palantir.atlasdb.transaction.impl;
 import com.google.common.collect.ImmutableMap;
 import com.palantir.atlasdb.keyvalue.api.KeyValueService;
 import com.palantir.atlasdb.keyvalue.api.TableReference;
-import com.palantir.atlasdb.transaction.api.TransactionReadInfo;
+import com.palantir.atlasdb.transaction.api.expectations.TransactionReadInfo;
 
 public interface TrackingKeyValueService extends KeyValueService {
     TransactionReadInfo getReadInfo();
