+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/transaction/impl/expectations/ExpectationsManager.java
@@ -16,7 +16,7 @@
 
 package com.palantir.atlasdb.transaction.impl.expectations;
 
-import com.palantir.atlasdb.transaction.api.ExpectationsAwareTransaction;
+import com.palantir.atlasdb.transaction.api.expectations.ExpectationsAwareTransaction;
 
 public interface ExpectationsManager extends AutoCloseable {
     void scheduleMetricsUpdate(long delayMillis);
